packages/python-rag-core/pyproject.toml
[project]
name = "python-rag-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Inesh-Reddy",email = "ineshreddy007@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.*"
dependencies = [
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "qdrant-client (>=1.16.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "python_rag_core", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
